+++ kythe/go/platform/tools/indexer_wrapper/BUILD
@@ -2,10 +2,7 @@ load("//tools:build_rules/shims.bzl", "go_binary")
 
 go_binary(
     name = "indexer_wrapper",
-    srcs = [
-        "local.go",
-        "main.go",
-    ],
+    srcs = ["main.go"],
     deps = [
         "//kythe/go/platform/analysis",
         "//kythe/go/platform/analysis/driver",
